Hi folks!

Due to very long execution times of our backtests, I made a performance profile. Almost half of the total execution time is used to serialize and then instantiate the `MarketBook` class. Using the lightweight option only partially reduces this performance overhead. I think unifying the caches and the betting data-structures could have a huge impact on overall performance. AFAIK this would also decrease the framework overhead in production, since the same cache is used. 

![Screenshot 2020-09-09 at 09 51 34](https://user-images.githubusercontent.com/9663231/92570541-52614c80-f282-11ea-88df-84a8e0b811f2.png)

### Prototype

I started to do a prototype implementation to validate this approach. I refactored parts of the `RunnerBook` cache to remove the overhead for serializing the `price`/`size` lists in the underlying [betfair](https://github.com/liampauling/betfair) project. This corresponds to the second back line marked in the profiling screenshot. This reduced the runtime for our benchmarks by a little over 10%, which verifies the assumptions made based on the profiling.

Since the `MarketBook` instances now partially hold data references to the cache, the historic stream used in the backtest had to be updated. pending orders are now processed before the cache is updated.

The current implementation is a WIP and not polished. After our initial findings we however wanted to coordinate with you, since the changes are quite extensive. I will create the corresponding PR to the betfair project and link it here.

Best Regards,
Felix