@liampauling 2020-09-09
Hi Felix,

Thanks for this, for context bflw and flumine has been created to try and balance speed with readability, flumine offers backtesting out of the box but doesn't necessarily mean it should be used for backtesting a large sample of markets. I would always recommend parsing out streaming data, taking advantage of a relational db or using a faster language when you are looking to backtest across thousands of markets. However always keen to make bflw faster and have discussed on the slack (have you joined?) regarding using C for some of the streaming cache.

Regarding speed I have thought about going down the route you have with pre caching to prevent duplicating serialization on every MarketBook creation but met in the middle with this [commit](https://github.com/liampauling/flumine/commit/64ae76f0cb5571a040ce9904febf35df733931c4). Do you have any further details on the speed improvements you are getting? Interested in the number of function calls in comparison to before.