@flxbe 2020-09-09
Hi Liam,

I will join the slack later today, thanks for the suggestion. I counted the number of functions calls for the market `1.172586390`:

```
Before: 22,647,565 function calls (22641270 primitive calls) in 16.252 seconds
Runtime before without profiling: 11.992 seconds
After: 25,061,453 function calls (25055173 primitive calls) in 15.994 seconds
Runtime after without profiling: 10.462 seconds
```

I understand and agree to the design decision to balance performance with speed. I however also think that it is possible to find a suitable structure to avoid the cache serialization and still have readable and maintainable code. Do you on principle agree to this hypothesis or do you think there are reasons that would make this change unfeasible?