@flxbe 2020-09-25
Hi!

I like your release plan. I also think a major release will be necessary since there will probably be breaking changes, even without deprecating python 3.5. What do you think about the future of the `lightweight` mode? After the refactor the serialization of the cache will be additional work, which seems odd for a supposedly faster mode. Effort-wise, adding a lightweight-compatible mode will be some overhead. I can however not assess the relative usage of the lightweight mode and the resistance of the community towards switching.

I would like to support the refactor, however there are currently a lot of projects on my desk. I can therefore not commit to any fixed timeline.