@liampauling 2021-01-11
Hi Felix,
Not sure if you have seen but I have been refactoring based on this and have released v2.12.0b0 which has the following speed improvements with no changes on use.

```
# lightweight (2.3x improvement)
#  master: 6.860s (100%|██████████| 10/10 [00:06<00:00,  1.46it/s])
#  2.12.0: 2.998s (100%|██████████| 10/10 [00:02<00:00,  3.34it/s])
# Non lightweight (3.4x improvement)
#  master: 36.005s (100%|██████████| 10/10 [00:35<00:00,  3.60s/it])
#  2.12.0: 10.625s (100%|██████████| 10/10 [00:10<00:00,  1.06s/it])
# Non lightweight with flumine patching/listener/stream (2.8x improvement)
#  master: 13.896s (100%|██████████| 10/10 [00:13<00:00,  1.39s/it])
#  2.12.0: 5.005s (100%|██████████| 10/10 [00:04<00:00,  2.00it/s])
```

I will close this and the bflw PR but thanks for the push to get this improved!